(almost) all offset issues fixed!
git-svn-id: https://svn.code.sf.net/p/doomlegacy/svn/legacy/trunk@406 5da24444-3c17-0410-b9f4-a1bbb4243164

+++ video/hardware/hwr_render.cpp
@@ -646,7 +646,7 @@ void modelpres_t::Project(Actor *p)
 {
   //CONS_Printf("modelpres_t::Project: Not yet implemented\n");
 }
-void PatchTexture::HWR_Draw(int x, int y, int flags)
+void PatchTexture::HWR_Draw(float x, float y, int flags)
 {
   static Geometry *geo = 0;
   static State *state = 0;
@@ -672,7 +672,7 @@ void PatchTexture::HWR_Draw(int x, int y, int flags)
   state->Apply();
   geo->Draw();
 }
-void LumpTexture::HWR_Draw(int x, int y, int flags)
+void LumpTexture::HWR_Draw(float x, float y, int flags)
 {
   //CONS_Printf("LumpTexture::HWR_Draw: Not yet implemented\n");
 }
